+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.2.0
+ * Added Chart type to allow for easier interactive charts.
+ * Added examples for the Chart type.
+ * Extended backend for allowing the modificaiton of drawing area scale and offset.
+ * Improved documentation.
+
 ## 0.1.1
  * Made 3d plot example interactive
  * Added demo GIF to readme